Move `@source` validations into one place and improve some error messages
